backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
pydantic==2.5.0

# Environment and HTTP
python-dotenv==1.0.0
requests==2.31.0

# Cache
redis==5.0.1

# Scheduling
APScheduler==3.10.4

# Testing
pytest==7.4.3
httpx==0.25.2